+++ src/components/media/media.js
@@ -2,7 +2,6 @@ import React from 'react';
 import classnames from 'classnames';
 
 import MediaItem from './components/media-item';
-import MediaContent from './components/media-content';
 
 import Element from '../element';
 
@@ -13,8 +12,6 @@ const Media = ({ children, className, ...props }) => (
 );
 
 Media.Item = MediaItem;
-
-Media.Content = MediaContent;
 
 Media.propTypes = {};